+++ CHANGELOG.md
@@ -20,6 +20,7 @@ All notable changes to this project starting with the 0.6.0 release will be docu
 - `site backups get --latest` bug wherein it was returning the oldest backup, rather than the most recent. (#770)
 - `sites aliases` will no longer tell you you have no sites when none of your domains include 'pantheon.io'. (#782)
 - `site tags add` now searches for existing tags before adding another. (#771)
+- `site redis clear` undefined-variable error has been fixed. (#799)
 
 ## [0.10.0] - 2015-12-15
 ### Added